# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Ongo Microservice Exercise",
        "description": "This is a Backend Challenge for Ongo",
        "contact": {
            "name": "Randy Arbolaez",
            "email": "randyarbolaez@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "tags": [{
                "name": "Auth",
                "description": "Authenicate"
            },
            {
                "name": "Admin",
                "description": "Admin Information"
            },
            {
                "name": "Task",
                "description": "Task Functions"
            }
        ],
        "servers": [{
            "url": "http://localhost:3000/api",
            "description": "Development server"
        }]
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": ["Auth"],
                "description": "Sign up for an account",
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["username", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sign up for an account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "username": {
                                            "type": "string"
                                        },
                                        "password": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/authenticate": {
            "post": {
                "tags": ["Auth"],
                "description": "Authenticate(login) to account",
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["username", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticate(login) to account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "username": {
                                            "type": "string"
                                        },
                                        "password": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": ["Admin"],
                "description": "get every user",
                "operationId": "getAllUsers",
                "security": [{
                    "bearerAuth": []
                }],
                "responses": {
                    "responses": {
                        "200": {
                            "description": "Get every user.",
                            "content": {
                                "application/json": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/delete/{id}": {
            "delete": {
                "tags": ["Admin"],
                "description": "Delete user and their tasks.",
                "operationId": "deleteAUserAndTheirTasks",
                "security": [{
                    "bearerAuth": []
                }],
                "responses": {
                    "200": {
                        "description": "User has been deleted.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "id of user to delete",
                    "required": true,
                    "schema": {
                        "items": {
                            "type": "string"
                        }
                    },
                    "style": "simple"
                }]
            }
        },
...
    }
}